+++ src/main/java/com/tokyovending/TokyoVending/services/VendingMachineService.java
@@ -1,8 +1,13 @@
 package com.tokyovending.TokyoVending.services;
 
 import com.tokyovending.TokyoVending.dtos.VendingMachineDto;
+import com.tokyovending.TokyoVending.models.Product;
+import com.tokyovending.TokyoVending.repositories.ProductRepository;
+import com.tokyovending.TokyoVending.exceptions.RecordNotFoundException;
+import com.tokyovending.TokyoVending.exceptions.EntityNotFoundException;
 import com.tokyovending.TokyoVending.models.VendingMachine;
 import com.tokyovending.TokyoVending.repositories.VendingMachineRepository;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,24 +17,24 @@ import java.util.stream.Collectors;
 @Service
 public class VendingMachineService {
 
-    private final VendingMachineRepository vendingMachineRepository;
-
     @Autowired
-    public VendingMachineService(VendingMachineRepository vendingMachineRepository) {
-        this.vendingMachineRepository = vendingMachineRepository;
-    }
+    private VendingMachineRepository vendingMachineRepository;
+    @Autowired
+    private ProductRepository productRepository;
+
 
     public List<VendingMachineDto> getAllVendingMachines() {
         List<VendingMachine> vendingMachines = vendingMachineRepository.findAll();
         return vendingMachines.stream().map(this::convertToDto).collect(Collectors.toList());
     }
 
+    @Transactional
     public VendingMachineDto getVendingMachineById(Long id) {
-        VendingMachine vendingMachine = vendingMachineRepository.findById(id).orElse(null);
-        if (vendingMachine != null) {
-            return convertToDto(vendingMachine);
-        }
-        return null;
+        VendingMachine vendingMachine = vendingMachineRepository.findById(id).orElseThrow(() ->
+                new RecordNotFoundException("VendingMachine with ID " + id + " not found.")
+        );
+        vendingMachine.getProducts().size();
+        return convertToDto(vendingMachine);
     }
 
     public VendingMachineDto createVendingMachine(VendingMachineDto vendingMachineDto) {
@@ -39,16 +44,18 @@ public class VendingMachineService {
     }
 
     public VendingMachineDto updateVendingMachine(Long id, VendingMachineDto vendingMachineDto) {
-        VendingMachine existingVendingMachine = vendingMachineRepository.findById(id).orElse(null);
-        if (existingVendingMachine != null) {
-            updateEntityFromDto(existingVendingMachine, vendingMachineDto);
-            VendingMachine updatedVendingMachine = vendingMachineRepository.save(existingVendingMachine);
-            return convertToDto(updatedVendingMachine);
-        }
-        return null;
+        VendingMachine existingVendingMachine = vendingMachineRepository.findById(id).orElseThrow(() ->
+                new RecordNotFoundException("VendingMachine with ID " + id + " not found.")
+        );
+        updateEntityFromDto(existingVendingMachine, vendingMachineDto);
+        VendingMachine updatedVendingMachine = vendingMachineRepository.save(existingVendingMachine);
+        return convertToDto(updatedVendingMachine);
     }
 
     public void deleteVendingMachine(Long id) {
+        if (!vendingMachineRepository.existsById(id)) {
+            throw new RecordNotFoundException("VendingMachine with ID " + id + " not found.");
+        }
         vendingMachineRepository.deleteById(id);
     }
 
@@ -56,6 +63,8 @@ public class VendingMachineService {
         VendingMachineDto vendingMachineDto = new VendingMachineDto();
         vendingMachineDto.setId(vendingMachine.getId());
         vendingMachineDto.setLocation(vendingMachine.getLocation());
+        vendingMachineDto.setOpen(vendingMachine.isOpen());
+        vendingMachineDto.setProducts(vendingMachine.getProducts());
         return vendingMachineDto;
     }
 
@@ -67,7 +76,22 @@ public class VendingMachineService {
 
     private void updateEntityFromDto(VendingMachine vendingMachine, VendingMachineDto vendingMachineDto) {
         vendingMachine.setLocation(vendingMachineDto.getLocation());
+        vendingMachine.setOpen(vendingMachineDto.isOpen());
     }
+
+    public VendingMachineDto addProductToVendingMachine(Long vmId, Long productId) {
+        VendingMachine vendingMachine = vendingMachineRepository.findById(vmId)
+                .orElseThrow(() -> new EntityNotFoundException("VendingMachine not found"));
+        Product product = productRepository.findById(productId)
+                .orElseThrow(() -> new EntityNotFoundException("Product not found"));
+
+        product.setVendingMachine(vendingMachine);
+        vendingMachine.getProducts().add(product);
+
+        VendingMachine updatedVendingMachine = vendingMachineRepository.save(vendingMachine);
+        return convertToDto(updatedVendingMachine);
+    }
+
 }
